Fixed bug at PDBIO.set_structure()
It always tries to access a chain object whose id is "A". However, if one tries to save a Residue/Atom whose chain id is not "A", it raises a KeyError exception because the recently created chain (line 79) had its id updated to pdb_object.parent.id at line 83 (when pdb_object is a residue) or 92 (when pdb_object is an atom).