Iterative implementation for newickize
`newickize` can [overflow the python stack for very deep trees](https://github.com/nextstrain/augur/issues/328).  This is an alternate implementation that uses a deque (as a stack) to store the tree data.  Although this is the same O(N) in terms of storage, the amount of data stored per function call depth in python is significantly higher.