SnapGene: Support features wrapping the origin.
A SnapGene file containing a circular sequence may contain feature
range specifications that wrap the beginning/end of the sequence,
e.g. "3925-2" for a feature starting at position 3925 near the end
of the sequence and ending at position 2. This case must be
recognized and represented as a CompoundLocation in the SeqRecord
object.