Test Bio.Phylo.write with NeXML and filename
This is to cover the code path where Bio.Phylo
has to open the output file and pick the
appropriate mode (ought to be binary), and
verify the previous commit restores the old
behaviour.

This does not actually fix Bio.Phylo to open
all XML files for reading or writing in binary
mode.