default phyml test to single threaded
By default the phyml test runs a MPI variant of the program that makes
the assumption of having access to an X server.  Since it may not be
the case on build servers, this outputs the message "No protocol
specified" in standard error, triggering a failure of the test.

This patch sets the variable PHYMLCPUS, if it is not in use, in order
to run phyml tests with the single threaded variant.  It is still
possible to override by setting with the environment variable if one
wishes to test explicitly the parallel case in the future.