SeqIO/SnapGene: Store feature name in specific qualifier
* SnapGene: Store feature name in specific qualifier (#3286).

Features in SnapGene files may carry a SnapGene-specific "name"
attribute containing a different value from the standard "label"
qualifier. If so, we store that name in a dedicated qualifier so
that it can be accessed by client code.

Thanks to GitHub user @nicolazilio for the bug report.

* SnapGene: Update test file.

Update the SnapGene sample files to test different combinations
of "name" and "label" feature attributes.