Dual license Bio/Phylo/Applications/_Raxml.py
Original author Eric Talevich and contributors since
tracked with version control (with one exception noted
below) have all agreed:

 - Adhemar Zerlotini (@azneto)
   https://github.com/biopython/biopython/pull/1412#issue-147372599
 - Andrey Raspopov (@Andrey-Raspopov)
   https://github.com/biopython/biopython/pull/2444
 - Chris Daley (@chebizarro)
   https://github.com/biopython/biopython/pull/2349
 - Christian Brueffer (@cbrueffer)
   https://github.com/biopython/biopython/issues/898#issuecomment-236922857
 - Eric Talevich (@etal)
   https://github.com/biopython/biopython/issues/898#issuecomment-236949433
 - Peter Cock (@peterjc)
   https://github.com/biopython/biopython/issues/898#issue-167420613
 - Sergio Valqui (@svalqui)
   https://github.com/biopython/biopython/pull/1749
 - Sujan Dulal (@Dsujan)
   https://github.com/biopython/biopython/pull/2430

We have not had explicit permission from Jingping Li
(@Jingping) for the one line change applied in
https://github.com/biopython/biopython/pull/73 and
credited to them. The change was suggested by me, but
the recorded authorship was more to recognise their
work in identifying the problem and testing the solution.