preserves dbsource annotation in GenBank output (#3425)
* Test SeqIO.write doesn't destroy DBSOURCE

Using SeqIO.write was not writing the full content of the annotations.
Namely the issue reported in #920 was that `db_source` was available
when reading from a file that had a DBSOURCE line, but it was removed on
SeqIO.write.

* Write DBSOURCE if db_source present in annotations

* Add self to CONTRIB.rst

* Fix test related to type of db_source entry

The BioSQL loader appears to return the annotation for db_source as a
list wrapping the original string. This was breaking, so added a check
for the type and just access to first value from the list.

* Remove reference to db_source as being unwritten

The TODO above this line indicates that this is a hack because the
"comment", "references", and "db_source" are not being written. This
branch is explicitly writing the "db_source" but we may also be able to
remove the other values here.