replace accession by locus id in EMBL header line
If the SeqRecord/SeqFeature API is used to create an in-memory data structure of an, e.g. bacterial genome annotation for subsequent serialization via `SeqIO`, the `GenBankWriter` writes both the *locus id* and the *accession*, latter only if available.

However, the `EmblWriter` either writes the *accession* in both fields ignoring the *locus id*

This is an issue when annotating local genome assemblies for which no public *accession* is available - but an unique *locus id* is of course required to identify the records within a multi EMBL/Genbank file.