Fix order of SnapGene join features on reverse strand (#3662)
* SnapGene: Fix ordering of reverse composite features.

The locations of the individual segments making up a feature must be
added in reverse order when the feature is on the minus strand.

* SnapGene: Skip "gap" segments for split features.

SnapGene uses a "gap" segment to represent the interval between two
parts of a split feature. Such segments must be skipped to build an
accurate CompoundLocation for the feature (otherwise the gap segments
are interpreted as being part of the feature).

* SnapGene: Check for multi-valued qualifiers and split features.

When testing the SnapGene parser, check that it correctly handles
qualifiers with multiple values and features split in several parts.

* Use a "parts" qualifier to represent named subfeatures.

SnapGene allows the different parts of a compound feature to have their
own name. This is not something we can directly represent in the
SeqRecord model, so if a feature part has a name, we translate it to
a distinct "misc_feature" in addition to the encompassing compound
feature.

If a multi-segment feature contains named segments, store the name of
those segments in a dedicated "parts" qualifier with the following
format:

  X:name_of_segment_X;Y:name_of_segment_Y

Use a 1-based index to identify segments, and for features on the
reverse strand, the first segment is the one appearing last.