adjust test_GraphicsBitmaps.py for newer renderpm (#3767)
Starting with reportlab 3.6, the RenderPM error message relative to
missing fonts does not start with "Can't setFont(" any more, but
"Error in setFont(" instead.  This breaks the test suite in case
the font is indeed missing, while it used to be skipped instead.
The main symptom is the test suite failing with error:

Traceback (most recent call last):
...
reportlab.graphics.utils.RenderPMError: Error in setFont('Times-Roman',10) missing the T1 files?
Originally <class 'TypeError'>: makeT1Font() argument 2 must be str, not None

A full log is available on Debian continuous integration platform [1].

[1]: https://ci.debian.net/data/autopkgtest/testing/arm64/p/python-biopython/16089190/log.gz

This patch adds a pattern to match for the new error message as well.