raise and error if an invalid Newick unquoted label name is found while parsing the tree
The issue is that the current regex expression that matches an
unquoted label returns several tokens for an invalid unquoted
label. Therefore, instead of raising an error, the function ends
up setting the node's name as the last token in the regex group
that matched the unquoted label name expression.

The solution I have come up with is to return the entire unquoted
label string with whitespaces, new lines, etc. included, and do the
checking manually when _parse_tree encounters an unquoted label
token.