Add ruff to pre-commit config
The tool ruff implements a subset of flake8 (which due
to runtime we already have to skip on the continuous
integration service pre-commit.ci), but is extremely
fast. Adding this makes minimal impact to running
the pre-commit checks locally on changed files, or
on all the code as used in continuous integration.