Test on Python 3.11
Would like to phase out Python 3.7 support, but
currently PyPy still targets that version of the
Python language - so not yet?

Also now recommending Python 3.10 (was Python 3.9),
not suggesting Python 3.11 at this point as it is
still quite new and not all third packages will have
been update for it yet (e.g. providing wheels).