Reimplement Bio.SeqUtils.ProtParam.flexibility as per Vihinen 1994 paper
Correct the number of flexibility scores

For a sequence of length `L` and a window size `W`, the number of flexibility
scores we can calculate is `L - W + 1`, not `L - W` as in the current code.
For example, a sequence of length 9 should have 1 score, but it is '0' in
the current version.

The flexibility should have 144 values for the length 152 sequence in the test.
The last value is '0.9822380952380954' for 'L' in 'AILFLPLPV' which is the last
9 residues.

The full implementation of the flexibility predictor method in the Vihinen 1994
paper requires all the B_norm values in Table 1.

Use neighbor-correlated B_norm values

According to the original Karplus 1985 paper, the neighbor-correlated B_norm
values are only calculated for non-terminal residues and thus the minimum
sequence length is `window_size + 2`.  Valuable names are changed to reflect
the Table I better.