snapgene: Fix parsing of primers coordinates.
Primers' coordinates in the Primers packet of a SnapGene file are
shifted by -1: a binding site really located in region 20-50 will have
its coordinates written down as 19-49.

This shifting of coordinates is specific to primers. All other features
(including `primer_bind` features described in the Features packet) have
their coordinates written down as they are.

Therefore, we apply a +1 correction to the coordinates of primers read
from the Primers packet.