Allow string-like objects as filenames in SeqIO
Rather than forcing the filename parameters to SeqIO.index and
SeqIO.index_db to be of type str, allow them to be any object that
implements the __str__ method for itself (ie: doesn't use the default
object.__str__). This allows pathlib.Path objects to be passed to those
functions and bascially anything else you'd want to represent a path.
See issue #4325.