Reverse chronological sort DEPRECATION file
This was done by hand, using the oldest version
mentioned in each section (be that a being declared
obsolete, deprecated, a removal, or a major change).

Potentially some of the longer-lived modules like
Bio.Align and Bio.SeqUtils would benefit from their
sections being split up into older entries (at the
bottom of the file) and newer entries (towards the
top of the file)?