Align codonaligner (#4479)
Adds a new CodonAligner class to Bio.Align. This class has the same capabilities
as the codon aligner in Bio.codonalign, but instead of regular expressions, it uses
a dynamic programming algorithm based on the pairwise aligner in Bio.Align.

Also adds a mapall method to the Alignment class. When applied to a protein alignment,
with a list of nucleotide-to-protein alignments generated using the CodonAligner, it will
return a nucleotide alignment where codons are aligned to each other.

The analysis methods calculate_dn_ds, calculate_dn_ds_matrix, and mktest are
provided in a new submodule Bio.Align.analysis, acting on an Alignment object
containing a nucleotide alignment of codons.

I believe that with this PR, Bio.codonalign can be deprecated (to be done later in a
separate PR).

This PR also fixes issue #4466.

As an aside, the sorts the bibliography entries in the Tutorial.