Add input check for uniprot-xml SeqIO format
Add check that aborts parsing when the input does not have the expected
xml namespace.  This will cause the parser
to fail with a clear error message and prevent misleading error messages
like the OOM error in issue #biopython/biopython#4467

This fix uses the xml namespaces in the input to validate, since the
parser is already assuming the input has uniprot namespace.