Ignore more doc8 warnings
D000 Unknown target name: seems to give false positives
on citations, probably as they are in a separate file.

D001 Line too long: we ought to review this later, we
might be able to improve on the pandoc layout.

D002 Trailing whitespace: probably from doctests, which
is hard to avoid.