Full author list on PDF front page
This works, is readable, but not as pretty as I'd
like. We might use the maketitle setting to setup
the entire PDF title page ourselves?

Note that as far as I can see, the author field
has no effect on the HTML output, thus we now have
a duplicated author listing in the index.rst file,
which is not ideal.