Use absolute /docs/... for version switching
Previously with just API docs, all the HTML files
lived in the same directory, so a relative path
was fine. Now we have the API and Tutorial in
sister folders, with the main index one level up.

The current relative links only worked from the
main index level, not the API or Tutorial pages.

I can't see an easy way to have context specific
relative links.