Drop Chain type annotation to appease Sphinx
This change surprisingly was blocking using a more recent
version of Sphinx (our issue #4584), see:

https://github.com/sphinx-doc/sphinx/issues/11883

The *pair* of warnings about Chain being ambiguous had mislead me
to think the trigger was the two Bio.PDB.Chain.Chain class
references in the docstrings.

TODO: Reinstate the type annotation for this method.