Add cif-seqres-author and cif-atom-label seq parsers
See #4922.

* The `CifSeqresIterator` (`cif-seqres`) uses "label" IDs
  ([`_pdbx_poly_seq_scheme.asym_id`][_pdbx_poly_seq_scheme.asym_id]),
  which are assigned by the PDB, always start at "A" and move through the
  alphabet one letter at a time.

* The `CifAtomIterator` (`cif-atom`) uses "author" IDs, which are
  arbitrarily assigned by the structure authors (this is what you get in
  traditional fixed-width PDB files).

I agree that this mismatch is probably a bug, though it would be a
backwards incompatible change to fix it. If you don't care about the
author IDs, and you just need the chain IDs of the sequence from
`cif-seqres` to match the chain IDs of the atoms parsed by
`MMCIFParser`, you could pass `auth_chains=False` to `MMCIFParser` to
use label IDs in the structure as well.

To fix this, the `CifSeqresIterator` could be modified to use author IDs
by reading
[`_pdbx_poly_seq_scheme.pdb_strand_id`][_pdbx_poly_seq_scheme.pdb_strand_id].
Or the `CifAtomIterator` could pass `auth_chains=False` to `MMCIFParser`
to use label IDs everywhere.

[_pdbx_poly_seq_scheme.asym_id]: https://mmcif.wwpdb.org/dictionaries/mmcif_pdbx_v50.dic/Items/_pdbx_poly_seq_scheme.asym_id.html
[_pdbx_poly_seq_scheme.pdb_strand_id]: https://mmcif.wwpdb.org/dictionaries/mmcif_ma.dic/Items/_pdbx_poly_seq_scheme.pdb_strand_id.html