fix(SeqFeature): add missing decorator to Location.fromstring factory method
The fromstring method of the abstract Location class is missing a method decorator, which causes the first argument to be interpreted as self.
This doesn't lead to any problems when calling it from the class, but confuses tooling.

Note that the commit that added this (908ee298) speaks of a "fromstring class method", so taking this method in the instance context should be fine.

I agree to my contributions being dual licensed under the BSD 3-Clause and Biopython License
Signed-off-by: Fynn Freyer <fynn.freyer@googlemail.com>