Add more ways to translate nucleotide sequence with dashes
Currently one cannot translate sequences aligned (padded)
with dashes into a protein sequence as the in-built tests
are way too strict. Add a couple of arguments to the
translate() functions to overcome the road blocks.

Typically padding symbols for gaps are '-' but this default
can be changed on the fly.

With this change one can with respect_alignment=True translate
a padded alignment file in FASTA format so that '---' translates
to '-', 'AT-' translates to 'X'.

In an alternative mode with ignore_gaps=True it is possible
to ignore the padding dashes so that 'AT-GATG' translates to
'MM'.